wip(attestation): agent-change index review fixes complete (gate green)
